Re-arrange tests
