+++ glue/core/data_factories/__init__.py
@@ -1,7 +1,8 @@
 from __future__ import absolute_import, division, print_function
 
 from .helpers import *
-from .containers import *
+from .fits import *
+from .hdf5 import *
 from .gridded import *
 from .pandas import *
 from .excel import *
